chore: a script to deploy a pseudo memecoin locally
